requirements.txt
boto3==1.17.44
botocore==1.20.44
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jmespath==0.10.0
numpy==1.20.2
pandas==1.2.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
tqdm==4.59.0
urllib3==1.26.4
